+++ diffprof/tests/test_latin_hypercube.py
@@ -1,27 +1,14 @@
 """
 """
-import pytest
 import numpy as np
 from ..latin_hypercube import latin_hypercube, latin_hypercube_from_cov
-from ..latin_hypercube import uniform_random_hypercube, latin_hypercube_pydoe
-from ..latin_hypercube import HAS_PYDOE2
+from ..latin_hypercube import uniform_random_hypercube
 
 
 def verify_lhs_respects_bounds(box, xmins, xmaxs):
     for idim in range(len(xmins)):
         assert np.all(box[:, idim] >= xmins[idim])
         assert np.all(box[:, idim] <= xmaxs[idim])
-
-
-@pytest.mark.xfail
-def test_latin_hypercube_pydoe_respects_bounds():
-    xmins = (-3, -2, 0)
-    xmaxs = (2, 3, 5)
-    n_dim = len(xmins)
-    npts = 5000
-    if HAS_PYDOE2:
-        lhs_box = latin_hypercube_pydoe(xmins, xmaxs, n_dim, npts)
-        verify_lhs_respects_bounds(lhs_box, xmins, xmaxs)
 
 
 def test_uniform_random_hypercube_respects_bounds():
@@ -43,20 +30,6 @@ def test_uniform_random_hypercube_is_reproducible():
     lhs_box2 = uniform_random_hypercube(xmins, xmaxs, n_dim, npts, seed=2)
     assert np.allclose(lhs_box, lhs_box1)
     assert not np.allclose(lhs_box, lhs_box2)
-
-
-@pytest.mark.xfail
-def test_latin_hypercube_pydoe_is_reproducible():
-    xmins = (-3, -2, 0)
-    xmaxs = (2, 3, 5)
-    n_dim = len(xmins)
-    npts = 5000
-    if HAS_PYDOE2:
-        lhs_box = latin_hypercube_pydoe(xmins, xmaxs, n_dim, npts, seed=0)
-        lhs_box1 = latin_hypercube_pydoe(xmins, xmaxs, n_dim, npts, seed=0)
-        lhs_box2 = latin_hypercube_pydoe(xmins, xmaxs, n_dim, npts, seed=2)
-        assert np.allclose(lhs_box, lhs_box1)
-        assert not np.allclose(lhs_box, lhs_box2)
 
 
 def test_latin_hypercube_is_reproducible():
